# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TravelHub API - Comprehensive travel management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{tripId}/progress:
    get:
      operationId: getTripProgress
      tags: [trips]
      summary: Get document upload progress for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripProgress"
...
  /trips/{tripId}/parking:
    get:
      operationId: listParkings
      tags: [parking]
      summary: List parking reservations for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of parking reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parking"
    post:
      operationId: createParking
      tags: [parking]
      summary: Add a parking reservation to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParkingInput"
      responses:
        "201":
          description: Created parking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parking"
  /trips/{tripId}/parking/{parkingId}:
    patch:
      operationId: updateParking
      tags: [parking]
      summary: Update a parking reservation
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: parkingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParkingUpdate"
      responses:
        "200":
          description: Updated parking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parking"
    delete:
      operationId: deleteParking
      tags: [parking]
      summary: Delete a parking reservation
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: parkingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{tripId}/rentals:
    get:
      operationId: listRentals
      tags: [rentals]
      summary: List vehicle rentals for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of vehicle rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rental"
    post:
      operationId: createRental
      tags: [rentals]
      summary: Add a vehicle rental to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalInput"
      responses:
        "201":
          description: Created rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /trips/{tripId}/rentals/{rentalId}:
    patch:
      operationId: updateRental
      tags: [rentals]
      summary: Update a vehicle rental
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: rentalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalUpdate"
      responses:
        "200":
          description: Updated rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
    delete:
      operationId: deleteRental
      tags: [rentals]
      summary: Delete a vehicle rental
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: rentalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Trip:
      type: object
      required: [id, name, destination, startDate, endDate, status, coverImage, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        coverImage:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TripInput:
      type: object
      required: [name, destination, startDate, endDate]
      properties:
        name:
          type: string
          minLength: 1
        destination:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverImage:
          type: string
        notes:
          type: string
    TripUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverImage:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        notes:
          type: string
    TripProgress:
      type: object
      required: [tripId, totalModules, completedModules, percentComplete, moduleBreakdown]
      properties:
        tripId:
          type: integer
        totalModules:
          type: integer
        completedModules:
          type: integer
        percentComplete:
          type: integer
        moduleBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ModuleProgress"
    ModuleProgress:
      type: object
      required: [module, label, hasDocuments, documentCount]
      properties:
        module:
          type: string
        label:
          type: string
        hasDocuments:
          type: boolean
        documentCount:
          type: integer
...
    Parking:
      type: object
      required: [id, tripId, location, reservationCode, entryDate, exitDate, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        location:
          type: string
        reservationCode:
          type: string
        entryDate:
          type: string
          format: date-time
        exitDate:
          type: string
          format: date-time
        priceTotal:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ParkingInput:
      type: object
      required: [location, reservationCode, entryDate, exitDate]
      properties:
        location:
          type: string
          minLength: 1
        reservationCode:
          type: string
          minLength: 1
        entryDate:
          type: string
          format: date-time
        exitDate:
          type: string
          format: date-time
        priceTotal:
          type: number
        notes:
          type: string
    ParkingUpdate:
      type: object
      properties:
        location:
          type: string
        reservationCode:
          type: string
        entryDate:
          type: string
          format: date-time
        exitDate:
          type: string
          format: date-time
        priceTotal:
          type: number
        notes:
          type: string
    Rental:
      type: object
      required: [id, tripId, company, pickupLocation, pickupDate, returnDate, fuelPolicy, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        company:
          type: string
        pickupLocation:
          type: string
        returnLocation:
          type: ["string", "null"]
        pickupDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
        fuelPolicy:
          type: string
        vehicleType:
          type: ["string", "null"]
        confirmationCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RentalInput:
      type: object
      required: [company, pickupLocation, pickupDate, returnDate, fuelPolicy]
      properties:
        company:
          type: string
          minLength: 1
        pickupLocation:
          type: string
          minLength: 1
        returnLocation:
          type: string
        pickupDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
        fuelPolicy:
          type: string
        vehicleType:
          type: string
        confirmationCode:
          type: string
        notes:
          type: string
    RentalUpdate:
      type: object
      properties:
        company:
          type: string
        pickupLocation:
          type: string
        returnLocation:
          type: string
        pickupDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
        fuelPolicy:
          type: string
        vehicleType:
          type: string
        confirmationCode:
          type: string
        notes:
          type: string
...