# lib/api-zod/src/generated/types/parkingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelHub API - Comprehensive travel management platform
 * OpenAPI spec version: 0.1.0
 */

export interface ParkingInput {
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  reservationCode: string;
  entryDate: Date;
  exitDate: Date;
  priceTotal?: number;
  notes?: string;
}