# lib/api-zod/src/generated/types/rental.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelHub API - Comprehensive travel management platform
 * OpenAPI spec version: 0.1.0
 */

export interface Rental {
  id: number;
  tripId: number;
  company: string;
  pickupLocation: string;
  /** @nullable */
  returnLocation?: string | null;
  pickupDate: Date;
  returnDate: Date;
  fuelPolicy: string;
  /** @nullable */
  vehicleType?: string | null;
  /** @nullable */
  confirmationCode?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}