# lib/api-zod/src/generated/types/rentalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelHub API - Comprehensive travel management platform
 * OpenAPI spec version: 0.1.0
 */

export interface RentalInput {
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  pickupLocation: string;
  returnLocation?: string;
  pickupDate: Date;
  returnDate: Date;
  fuelPolicy: string;
  vehicleType?: string;
  confirmationCode?: string;
  notes?: string;
}